+++ Werk/WorkoutList/WorkOutListView.swift
@@ -53,7 +53,7 @@ struct WorkOutListView: View {
 struct WorkOutListView_Previews: PreviewProvider {
     static var previews: some View {
         WorkOutListView(viewModel: WorkoutListViewModel(workOuts: (0...9).map{_ in
-            WorkOutExercise.randomWorkout}))
+            WorkoutTemplate.randomWorkout}))
     }
 
 }
